Split: TestChemicalFormula
